Refactoring: move SharingElectorateInfoOptions to Registration. Prepare ServerUser for filtering data sent tot server
